+++ devices/edge-router.js
@@ -1,6 +1,7 @@
 const Q = require('q');
 const MQTT = require('mqtt');
 const { async } = require('q');
+const COFS = require('../lib/cofs')
 const GetPublicIp = require('public-ip').v4;
 const EventEmitter = require('events').EventEmitter;
 
@@ -51,31 +52,12 @@ module.exports = class extends EventEmitter {
         this.barcode = args.barcode;
         this.deviceId = args.deviceId;
 
+        this.cofs = new COFS()
+
+        this.fixedTransmit = setInterval(() => this.transmit(), this.txtime * 1000);
+        clearInterval(this.fixedTransmit)
+
         this.connect();
-    }
-
-    send = () => {
-        var deferred = Q.defer()
-        let index = 0
-        __routerStatus.map(o => {
-            index++
-            __router.publish({
-                'rtuId': __settings.deviceId,
-                'dataIn': o.dataIn,
-                'barcode': __settings.barcode,
-                'rtuDate': new Date().getTime(),
-                'moduleId': o.moduleId
-            });
-            if (index == __routerStatus.length) {
-                deferred.resolve({})
-            }
-        })
-
-        if (index == 0) {
-            deferred.resolve({})
-        }
-
-        return deferred.promise
     }
 
     updateDeviceInputsThenActionMapping(deviceId, inputs) {
@@ -98,47 +80,6 @@ module.exports = class extends EventEmitter {
             __router.mapping(deviceId, inputs)
         })
 
-    }
-
-    applyCOFSServer() {
-        var deferred = Q.defer()
-        let index = 0
-
-        __devices.map(device => {
-            index++
-            if (device.enabled == true && device.publish == true) {
-                device.io.map(io => {
-                    if (io.publish.enabled == true) {
-                        let moduleFound = __routerStatus.find(o => o.moduleId == io.publish.moduleId)
-                        if (!moduleFound) {
-                            __routerStatus.push({ moduleId: __routerStatus.length, dataIn: { ...dataIn } })
-                        }
-                        if (io.publish.key == 'digitalsIn' && (parseFloat(io.publish.bit) != -1)) {
-                            let digitalsIn = __routerStatus[io.publish.moduleId].dataIn[io.publish.key]
-                            if (digitalsIn & Math.pow(2, io.publish.bit) > 0) {
-                                digitalsIn = digitalsIn - Math.pow(2, io.publish.bit)
-                            }
-                            if (io.value == 1) {
-                                digitalsIn = digitalsIn + Math.pow(2, io.publish.bit)
-                            }
-
-                            __routerStatus[io.publish.moduleId].dataIn[io.publish.key] = digitalsIn
-                        } else {
-                            __routerStatus[io.publish.moduleId].dataIn[io.publish.key] = io.value
-                        }
-                    }
-                })
-            }
-            if (__devices.length == index) {
-                deferred.resolve({})
-            }
-        })
-
-        if (index == 0) {
-            deferred.resolve({})
-        }
-
-        return deferred.promise
     }
 
     async mapping(deviceId, inputs) {
@@ -241,8 +182,8 @@ module.exports = class extends EventEmitter {
             })
         }, Promise.resolve())
         .then(async ()=>{
-            await this.applyCOFSServer()
-            await this.send()
+            await this.cofs.applyCOFSServer()
+            await this.cofs.send()
             deferred.resolve({})
         })
 
@@ -277,6 +218,7 @@ module.exports = class extends EventEmitter {
         });
 
         this.mqtt.on('close', () => {
+            clearInterval(this.fixedTransmit)
             this.status = 'disconnected';
             __logger.error('Edge Router - Socket closed!');
         });
@@ -316,9 +258,12 @@ module.exports = class extends EventEmitter {
 
             __logger.info('Edge Router - Starting transmit loop!');
 
-            // this.transmit();
+            setTimeout(()=>{
+                this.transmit();
+            },10000)
 
-            setInterval(() => this.transmit(), this.txtime * 1000);
+            clearInterval(this.fixedTransmit)
+            this.fixedTransmit
         });
 
         this.mqtt.on('message', async (topic, message) => {
